--- cps.py
@@ -1,57 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-#  This file is part of the Calibre-Web (https://github.com/janeczku/calibre-web)
-#    Copyright (C) 2022 OzzieIsaacs
-#
-#  This program is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#
-#  This program is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-import os
-import sys
-
-
-# Add local path to sys.path, so we can import cps
-path = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, path)
-
-from cps.setup_manager import is_first_run, run_interactive_setup
-from cps.main import main
-
-
-def hide_console_windows():
-    import ctypes
-
-    kernel32 = ctypes.WinDLL('kernel32')
-    user32 = ctypes.WinDLL('user32')
-
-    SW_HIDE = 0
-
-    hWnd = kernel32.GetConsoleWindow()
-    if hWnd:
-        user32.ShowWindow(hWnd, SW_HIDE)
-
-
-if __name__ == '__main__':
-    # Check if first-run setup is needed
-    if is_first_run():
-        print("\n🚀 First-run setup detected. Starting configuration wizard...\n")
-        if not run_interactive_setup():
-            print("\n❌ Setup cancelled or failed. Exiting.")
-            sys.exit(1)
-    
-    if os.name == "nt":
-        hide_console_windows()
-    main()
-
-
