--- pyproject.toml
@@ -1,47 +1,0 @@
-[project]
-name = "getmyebook-web"
-version = "0.1.0"
-
-dependencies = [
-    "apscheduler>=3.6.3,<3.11.0",
-    "babel>=1.3,<3.0",
-    "flask-babel>=0.11.1,<4.1.0",
-    "flask-principal>=0.3.2,<0.5.1",
-    "flask==2.3.3",
-    "iso-639>=0.4.5,<0.5.0",
-    "pypdf>=3.15.6,<5.1.0",
-    "pytz>=2016.10",
-    "requests>=2.28.0,<2.32.0",
-    "sqlalchemy>=2.0.0,<2.1.0",
-    "tornado>=6.3,<6.5",
-    "wand>=0.4.4,<0.7.0",
-    "unidecode>=0.4.19,<1.4.0",
-    "lxml>=4.9.1,<5.3.0",
-    "flask-wtf>=0.14.2,<1.3.0",
-    "chardet>=3.0.0,<5.3.0",
-    "netifaces-plus>0.12.0,<0.13.0",
-    "urllib3>=1.22,<2.0",
-    "flask-limiter>=2.3.0,<3.9.0",
-    "regex>=2022.3.2,<2024.6.25",
-    "bleach>=6.0.0,<6.2.0",
-    "python-magic>=0.4.27,<0.5.0",
-    "flask-httpauth>=4.4.0,<5.0.0",
-    "cryptography>=30.0.0,<44.0.0",
-    "dotenv==0.9.9",
-    "authlib==1.6.5",
-    "python-dotenv>=1.2.1",
-    "psycopg2-binary==2.9.11",
-    # Forum dependencies
-    "flask-sqlalchemy>=3.0.0,<4.0.0",
-    "flask-migrate>=4.0.0,<5.0.0",
-    "flask-bcrypt>=1.0.0,<2.0.0",
-    "flask-mail>=0.9.0,<1.0.0",
-    "flask-marshmallow>=1.0.0,<2.0.0",
-    "flask-login>=0.6.0,<1.0.0",
-    "marshmallow-sqlalchemy>=1.0.0,<2.0.0",
-    "python-slugify>=8.0.0,<9.0.0",
-    "marshmallow>=4.1.0",
-    "cachelib>=0.13.0,<0.14.0",
-    "alabaster>=1.0.0",
-    "boto3>=1.43.6",
-]
